[Refactor] MemberController 불필요한 매개변수 제거, Oauthsuccess핸들러 닉네임 부분 수정

See merge request s08-webmobile2-sub2/S08P12B111!26

+++ back/src/main/java/com/udteam/miristock/dto/CommentRequestDto.java
@@ -13,9 +13,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 public class CommentRequestDto {
 
-    private Long commentNo; // 댓글 번호
-
-    private Long articleNo;
+    private Integer commentNo; // 댓글 번호
+    private Integer articleNo;
     private String memberNickname; // 작성자 닉네임
     private String commentContent; // 내용
     private LocalDateTime commentDate; // 작성시간
